### 2022-07-31 #4631 @editorialbot
[PRE REVIEW]: metrica: an R package to evaluate prediction performance of regression and classification point-forecast models
**Submitting author:** @adriancorrendo (<a href="http://orcid.org/0000-0002-4172-289X">Adrian Correndo</a>)
**Repository:** https://github.com/adriancorrendo/metrica/
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v2.0.1
