### 2022-07-31 715289b
JOSS pre-rev-1: fix DOIs and affiliation index
