+++ paper.bib
@@ -13,7 +13,7 @@ volume = {192},
 pages = {103194},
 year = {2021},
 issn = {0308-521X},
-doi = {https://doi.org/10.1016/j.agsy.2021.103194},
+doi = {10.1016/j.agsy.2021.103194},
 url = {https://www.sciencedirect.com/science/article/pii/S0308521X21001475},
 author = {Adrian A. Correndo and Trevor J. Hefley and Dean P. Holzworth and Ignacio A. Ciampitti}
 }
@@ -124,7 +124,7 @@ author = {Adrian A. Correndo and Trevor J. Hefley and Dean P. Holzworth and Igna
     volume = {15},
     number = {4},
     pages = {14-19},
-    doi = {https://doi.org/10.1111/j.1740-9713.2018.01169.x},
+    doi = {10.1111/j.1740-9713.2018.01169.x},
     url = {https://rss.onlinelibrary.wiley.com/doi/abs/10.1111/j.1740-9713.2018.01169.x},
     eprint = {https://rss.onlinelibrary.wiley.com/doi/pdf/10.1111/j.1740-9713.2018.01169.x},
     abstract = {The story of a statistical programming language that became a subcultural phenomenon. By Nick Thieme},
@@ -154,7 +154,7 @@ edition = {Third Edition},
 pages = {311-373},
 year = {2019},
 isbn = {978-0-12-811756-9},
-doi = {https://doi.org/10.1016/B978-0-12-811756-9.00009-5},
+doi = {10.1016/B978-0-12-811756-9.00009-5},
 url = {https://www.sciencedirect.com/science/article/pii/B9780128117569000095},
 author = {Daniel Wallach and David Makowski and James W. Jones and François Brun},
 }
@@ -167,7 +167,7 @@ number = {2},
 pages = {225-247},
 year = {2006},
 issn = {0308-521X},
-doi = {https://doi.org/10.1016/j.agsy.2005.11.004},
+doi = {10.1016/j.agsy.2005.11.004},
 url = {https://www.sciencedirect.com/science/article/pii/S0308521X05002568},
 author = {Luis Orlindo Tedeschi},
 }
@@ -179,7 +179,7 @@ volume = {103},
 pages = {43-51},
 year = {2018},
 issn = {1364-8152},
-doi = {https://doi.org/10.1016/j.envsoft.2018.02.002},
+doi = {10.1016/j.envsoft.2018.02.002},
 url = {https://www.sciencedirect.com/science/article/pii/S1364815217311921},
 author = {Dean Holzworth and N.I. Huth and J. Fainges and H. Brown and E. Zurcher and R. Cichota and S. Verrall and N.I. Herrmann and B. Zheng and V. Snow},
 }
@@ -191,7 +191,7 @@ volume = {62},
 pages = {327-350},
 year = {2014},
 issn = {1364-8152},
-doi = {https://doi.org/10.1016/j.envsoft.2014.07.009},
+doi = {10.1016/j.envsoft.2014.07.009},
 url = {https://www.sciencedirect.com/science/article/pii/S1364815214002102},
[…]
