### 2022-07-31 #4631 @adriancorrendo
Thank you for starting the pre-review of the article!

I have just pushed a review to the JOSS branch at https://github.com/adriancorrendo/metrica/tree/JOSS fixing:

1) the missing and invalid DOIs in the "paper.bib" file https://github.com/openjournals/joss-reviews/issues/4631#issuecomment-1200462451.
2) fixing the affiliation index of L. Bastos in the "paper.md" file that generated the error compiling the pdf https://github.com/openjournals/joss-reviews/issues/4631#issuecomment-1200462527.
3) Also attached the draft pdf compiled with the draft-pdf.yaml https://github.com/adriancorrendo/metrica/blob/JOSS/paper.pdf

Sincerely,
ADRIAN
