### 2022-08-03 #4631 @neerajdhanraj
Yeah, I would like to review this article.

Best regards,

Neeraj Dhanraj Bokde, Ph.D.
https://neerajbokde.in/

On Wed, 3 Aug 2022, 12:23 pm Øystein Sørensen, ***@***.***>
wrote:

> 👋 @kauedesousa <https://github.com/kauedesousa> @tscheypidi
> <https://github.com/tscheypidi> @neerajdhanraj
> <https://github.com/neerajdhanraj> would any of you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
