### 2022-08-05 1a71b1c
fixing "doi:10.14358/PERS.72.7.823" in "paper.bib "
