+++ paper.bib
@@ -287,7 +287,7 @@ pages = {823-833},
 itemtype = {ARTICLE},
 issn = {0099-1112},
 url = {https://www.ingentaconnect.com/content/asprs/pers/2006/00000072/00000007/art00006},
-doi = {doi:10.14358/PERS.72.7.823},
+doi = {10.14358/PERS.72.7.823},
 author = {Ji, Lei and Gallo, Kevin},
 }
 
