### 2022-08-05 #4655 @adriancorrendo
Thank you!
I just updated the branch containing the paper https://github.com/adriancorrendo/metrica/tree/JOSS ,
fixing the INVALID DOI within the "paper.bib" under the following commit:
https://github.com/adriancorrendo/metrica/commit/1a71b1cd1ee96a9cb7610725861e5bc8d52a9bed

Best regards,

ADRIAN
