### 2022-10-04 #4655 @osorensen
👋  @wiljnich @simonpcouch @Emma-k-ward, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
