### 2022-10-04 #4655 @osorensen
Thanks @simonpcouch :-)  You can get started by issuing the command @editorialbot generate my checklist.
