### 2022-10-06 adriancorrendo/metrica#28 @simonpcouch
JOSS reviewer comments, submission 4655
Hey there!

This is part of my review for your submission to JOSS at openjournals/joss-reviews#4655.

I really enjoyed checking out this package! The package is principled and very well documented. Particularly well-scoped—I appreciate that you defer to existing data structures for function output, e.g. lists and data frames, and existing tools for existing operations in vignettes, e.g. `purrr::group_map` for applications to list-columns.

A few comments:

> **Data sharing:** If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.

For the most part, this is true, though the data included [in some vignettes](https://adriancorrendo.github.io/metrica/articles/apsim_open.html) under the `testthat` directory is not immediately available for users of the software:

``` r
soybean.out <- metrica::import_apsim_out(filepath = "../tests/testthat/examples/soybean.out")

#...

soybean.db <- metrica::import_apsim_db(filename = "soybean.example.db", folder = "../tests/testthat/examples/")
```

I was able to read these in by pulling the source locally and adjusting the filepaths. These either ought to be included as data with the package, saved to file in the vignette, and then read in with the `metrica::import_*` functions, or simply read in as files from a GitHub `raw` URL. The `file.path()` function can be used to create OS-independent filepaths.

> **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

You write in your paper "the first step is to install R (>= 4.2.0)," though your listed needed R version in DESCRIPTION is 2.1.0. These ought to be consistent, if included in the paper.

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I haven't been able to track down a CONTRIBUTING file, code of conduct, README section on contributing, or GitHub issue/PR template. I would feel confident checking off this box with any one of these present!

> **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

The help-files are really well-written. Many of the them refer to [these vignettes](https://adriancorrendo.github.io/metrica/articles/available_metrics_classification.html) giving formulas and higher-level descriptions, though many of the terms in the formulas are not documented. `B^2`, for instance, is ambiguous in context, and the `TP`/`TN`/`FP`/etc notations are not self-explanatory.

-------

As for the paper,

> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

Your bullet points beginning at line 60 are generally informative for describing the unique advantages of your package.

In the paper, you write:

> -   the most extensive collection of prediction performance metrics for regression and classification models up to date.

At first glance, this doesn't seem to be true given the packages you've cited. Loosely:

``` r
library(cvms)
length(formalArgs(gaussian_metrics)) +
length(formalArgs(binomial_metrics)) +
length(formalArgs(multinomial_metrics))
#> [1] 84
```

or

https://github.com/tidymodels/yardstick/blob/main/NAMESPACE#L3-L182

Metric calculation modules/libraries from non-R ecosystems, as well, can be [quite extensive](https://scikit-learn.org/stable/modules/model_evaluation.html).

...which is okay! I'd just recommend removing that line from the paper, or clarifying it with some qualifiers.

> **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The paper is generally well-written!

The "Package features" section is quite long, and includes elements that I haven't seen in other accepted JOSS papers. I'd recommend cutting the "Package features" section from "System requirements and installation" on, currently lines 86 through 109.

I came across a few grammatical errors while reviewing, and would like to see another round of copy edits before recommending acceptance. Would be glad to put in a PR with copy edits, if appreciated!

> **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

Many software names in these entries are not case protected, resulting in the first letters to be capitalized / made lowercase when they ought not to be. e.g. "Cvms: ..." ought to read `{cvms}:`... in the LaTeX.
