### 2022-10-07 #4655 @wiljnich
@adriancorrendo I've completed my initial review and request only minor revisions, all of which I feel are sufficiently addressed by @simonpcouch's excellent and thorough [issue #28](https://github.com/adriancorrendo/metrica/issues/28). Thank you for your hard work and forthcoming updates.
