+++ paper.bib
@@ -1,5 +1,5 @@
 @manual{metrica_manual,
-    title = {metrica: Prediction Performance Metrics},
+    title = {{metrica}: Prediction Performance Metrics},
     author = {Adrian A. Correndo and Luiz H. {Moro Rosso} and Rai Schwalbert and Carlos Hernandez and Leonardo M. Bastos and Luciana Nieto and Dean Holzworth and Ignacio A. Ciampitti},
     year = {2022},
     note = {R package version 2.0.1},
@@ -20,7 +20,7 @@ author = {Adrian A. Correndo and Trevor J. Hefley and Dean P. Holzworth and Igna
 
 
 @manual{yardstick_manual,
-    title = {yardstick: Tidy Characterizations of Model Performance},
+    title = {{yardstick}: Tidy Characterizations of Model Performance},
     author = {Max Kuhn and Davis Vaughan},
     year = {2022},
     note = {R package version 1.0.0},
@@ -35,7 +35,7 @@ author = {Adrian A. Correndo and Trevor J. Hefley and Dean P. Holzworth and Igna
   }
 
 @Manual{caret_manual,
-    title = {caret: Classification and Regression Training},
+    title = {{caret}: Classification and Regression Training},
     author = {Max Kuhn},
     year = {2022},
     note = {R package version 6.0-92},
@@ -56,7 +56,7 @@ author = {Adrian A. Correndo and Trevor J. Hefley and Dean P. Holzworth and Igna
 
 @Book{ggplot_book,
     author = {Hadley Wickham},
-    title = {ggplot2: Elegant Graphics for Data Analysis},
+    title = {{ggplot2}: Elegant Graphics for Data Analysis},
     publisher = {Springer-Verlag New York},
     year = {2016},
     isbn = {978-3-319-24277-4},
@@ -213,24 +213,16 @@ author = {Dean P. Holzworth and Neil I. Huth and Peter G. deVoil and Eric J. Zur
 }
 
 @manual{cvms_package,
-    title = {cvms: Cross-Validation for Model Selection},
+    title = {{cvms}: Cross-Validation for Model Selection},
     author = {Ludvig Renbo Olsen and Hugh Benjamin Zachariae},
     year = {2021},
     note = {R package version 1.3.3},
     url = {https://CRAN.R-project.org/package=cvms},
   }
 
-
-@manual{cvms_package,
-    title = {cvms: Cross-Validation for Model Selection},
-    author = {Ludvig Renbo Olsen and Hugh Benjamin Zachariae},
-    year = {2021},
-    note = {R package version 1.3.3},
-    url = {https://CRAN.R-project.org/package=cvms},
-  }
 
 @manual{scoringutils_package,
-    title = {scoringutils: Utilities for Scoring and Assessing
+    title = {{scoringutils}: Utilities for Scoring and Assessing
       Predictions},
     author = {Nikos I. Bosse and Hugo Gruson and Sebastian Funk and
       {EpiForecasts} and Sam Abbott},
@@ -368,7 +360,7 @@ author = {J.M. Yang and J.Y. Yang and S. Liu and G. Hoogenboom},
 
 @article{Willmott_1981,
 author = { Cort J.   Willmott },
-title = {ON THE VALIDATION OF MODELS},
+title = {On the validation of models},
 journal = {Physical Geography},
 volume = {2},
 number = {2},
@@ -474,7 +466,6 @@ author = {Harald Kling and Martin Fuchs and Maria Paulin},
 }
 
 @article{ferri2009,
[…]
