+++ paper.md
@@ -58,7 +58,7 @@ Finally, it is noteworthy that in the area of agricultural sciences, although po
 
 # Package features
 
-For regression models, `metrica` includes four plotting functions (scatter, tiles, density, & Bland-Altman plots) using `ggplot2` [@ggplot_book], and 48 prediction performance metrics. For classification models (two-class or multi-class), it includes one function to visualize a confusion matrix, and 27 functions of prediction scores. The full list of metrics with description, formula, and literature sources is presented in the package documentation at:
+For regression models, `metrica` includes four plotting functions (scatter, tiles, density, & Bland-Altman plots) using `ggplot2` [@ggplot_book], and 48 prediction performance metrics. For classification models (two-class or multi-class), it includes one function to visualize a confusion matrix, and 26 functions of prediction scores. The full list of metrics with description, formula, and literature sources is presented in the package documentation at:
 
 - Regression metrics: https://adriancorrendo.github.io/metrica/articles/available_metrics_regression.html.
 
@@ -66,7 +66,7 @@ For regression models, `metrica` includes four plotting functions (scatter, tile
 
 To extent of our knowledge, *metrica* covers several functions not supported, or partially supported by similar R packages (or components) designed for model evaluation such as `yardstick` [@yardstick_manual] from `tidymodels` [@tidymodels_manual], the measuring performance components from `caret` [@caret_manual] or `mlr3` [@mlr3_paper], `Metrics` [@Metrics_manual], `hydroGOF` [@hydroGOF_manual], `cvms` [@cvms_package], `scoringutils`[@scoringutils_package], or `performance` [@performance_paper]. Unique features include:
 
-- the most extensive collection of prediction performance metrics for regression and classification models up to date.
+- one of the most extensive collections of prediction performance metrics for regression and classification models in R.
 
 - working under both vectorized (calling variables with $) or tabulated forms [@tidyverse_paper].
 
@@ -82,166 +82,27 @@ To extent of our knowledge, *metrica* covers several functions not supported, or
 
 - importing files from APSIM Classic with `import_apsim_out()`), and from APSIM Next Generation with the `import_apsim_db()` function.
 
-## System requirements and installation
-
-Since *metrica* operates within R, the first step is to install R (>= 4.2.0). To install and load the package:
-
-```r
-# Stable version (CRAN)
-install.packages("metrica")
-
-# Development version (GitHub)
-devtools::install_github("adriancorrendo/metrica")
-
-# Load
-library(metrica)
-
-```
-
 ## Using the functions
 
 There are two core arguments to all *metrica* functions: (i) `obs`(Oi; observed, a.k.a. actual, measured, truth, target, label), and (ii) `pred` (Pi; predicted, a.k.a. simulated, fitted, modeled, estimate) values.
-For regression, specific functions require defining the axis `orientation` (*e.g.* predicted vs. observed -PO- or observed vs. predicted -OP-).
 
-For two-class models, the `pos_level` argument serves to indicate the alphanumeric order of the “positive level”. Following most two-class denominations as c(0,1), c(“Negative”, “Positive”), and c(“FALSE”, “TRUE”), the default `pos_level = 2` (1, “Positive”, “TRUE”). However, we recognize other cases as possible (e.g. c(“Crop”, “NoCrop”)), for which the user needs to specify `pos_level = 1`. For multi-class classification, some functions present the `atom` argument (TRUE / FALSE), which controls the output to be an overall average estimate across all classes (default), or class-wise.
[…]
