+++ paper.bib
@@ -181,7 +181,7 @@ year = {2018},
 issn = {1364-8152},
 doi = {10.1016/j.envsoft.2018.02.002},
 url = {https://www.sciencedirect.com/science/article/pii/S1364815217311921},
-author = {Dean Holzworth and N.I. Huth and J. Fainges and H. Brown and E. Zurcher and R. Cichota and S. Verrall and N.I. Herrmann and B. Zheng and V. Snow},
+author = {D. Holzworth and N. Huth and J. Fainges and H. Brown and E. Zurcher and R. Cichota and S. Verrall and N. Herrmann and B. Zheng and V. Snow},
 }
 
 @article{APSIM_classic,
@@ -193,7 +193,7 @@ year = {2014},
 issn = {1364-8152},
 doi = {10.1016/j.envsoft.2014.07.009},
 url = {https://www.sciencedirect.com/science/article/pii/S1364815214002102},
-author = {Dean P. Holzworth and Neil I. Huth and Peter G. deVoil and Eric J. Zurcher and Neville I. Herrmann and Greg McLean and Karine Chenu and Erik J. {van Oosterom} and Val Snow and Chris Murphy and Andrew D. Moore and Hamish Brown and Jeremy P.M. Whish and Shaun Verrall and Justin Fainges and Lindsay W. Bell and Allan S. Peake and Perry L. Poulton and Zvi Hochman and Peter J. Thorburn and Donald S. Gaydon and Neal P. Dalgliesh and Daniel Rodriguez and Howard Cox and Scott Chapman and Alastair Doherty and Edmar Teixeira and Joanna Sharp and Rogerio Cichota and Iris Vogeler and Frank Y. Li and Enli Wang and Graeme L. Hammer and Michael J. Robertson and John P. Dimes and Anthony M. Whitbread and James Hunt and Harm {van Rees} and Tim McClelland and Peter S. Carberry and John N.G. Hargreaves and Neil MacLeod and Cam McDonald and Justin Harsdorf and Sara Wedgwood and Brian A. Keating},
+author = {D. Holzworth and N. Huth and P. deVoil and E. Zurcher and N. Herrmann and G. McLean and K. Chenu and E. {van Oosterom} and V. Snow and C. Murphy and A. Moore and H. Brown and J. Whish and S. Verrall and J. Fainges and L. Bell and A. Peake and P. Poulton and Z. Hochman and P. Thorburn and D. Gaydon and N. Dalgliesh and D. Rodriguez and H. Cox and S. Chapman and A. Doherty and E. Teixeira and J. Sharp and R. Cichota and I. Vogeler and F. Li and E. Wang and G. Hammer and M. Robertson and J. Dimes and A. Whitbread and J. Hunt and H. {van Rees} and T. McClelland and P. Carberry and J. Hargreaves and N. MacLeod and C. McDonald and J. Harsdorf and S. Wedgwood and B. Keating},
 }
 
 @article{Seibold_2022,
