+++ paper.bib
@@ -285,7 +285,7 @@ author = {Ji, Lei and Gallo, Kevin},
 
 @article{Duveiller_2016,
 title = {Revisiting the concept of a symmetric index of agreement for continuous datasets},
-author = {Duveiller, G., Fasbender, D., and Meroni, M.},
+author = {G. Duveiller and  D. Fasbender and M. Meroni},
 journal = {Scientific Reports},
 year = {2016},
 volume = {6},
