### 2022-11-02 #4655 @adriancorrendo
> @adriancorrendo, could you please give us an update on how long you expect it will take to address the issues raised by the reviewers? It's absolutely fine if things take time, but it would be nice to have an estimate.

Good morning, Dr. Sorensen!

How are you?

Thank you for reaching out. My apologies for the delay.
Actually, I am planning to submit the response today by COB. It just took me longer than expected to work on the changes to paper, updating the code and documentation (latest commit to the [JOSS-branch here](https://github.com/adriancorrendo/metrica/tree/JOSS)), and putting things in common with the coauthors.

I also have a quick question regarding the response: most of reviews are provided on
[this issue](https://github.com/adriancorrendo/metrica/issues/28) by @simonpcouch .

But where I am supposed to provide the point by point response: i) here, ii) on the [same issue](https://github.com/adriancorrendo/metrica/issues/28),  or iii) create a new issue with the answers?

Thank you very much for your time and patience!

Best regards,
ADRIAN
