### 2022-11-02 #4655 @osorensen
@adriancorrendo, please respond in the issue in the source repository https://github.com/adriancorrendo/metrica/issues/28

The additional comment by @simonpcouch can be answered in this thread.
