### 2022-11-02 #4655 @adriancorrendo
Dear @simonpcouch and @wiljnich
First of all, thank you very much for your compelling revision of the package!
It certainly served to improve important aspects of this project.

I will proceed to answer the most of the revisions on the corresponding [issue#28](https://github.com/adriancorrendo/metrica/issues/28) opened by @simonpcouch on the source repository.

Please, you may read below the answer related to the **Contribution and authorship**:

> > **Contribution and authorship:** Has the submitting author ([@adriancorrendo](https://github.com/adriancorrendo)) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> The submitting author has surely made major contributions to the software! The extent of contributions from the remaining six authors are not self-evident given the GitHub commit history, though I recognize that [indeed](https://joss.readthedocs.io/en/latest/review_criteria.html#authorship) "authorship is a complex topic with different practices in different communities."

- Answer:
Thank you very much for this comment. You are correct with the observation, most of the code has been uploaded by myself (@adriancorrrendo) but it is simply the manner we have found suitable in our case. There has been a considerable amount of discussions between all the coauthors about key concepts, metrics, examples, data, etc. to give the necessary structure to the package. Many times, coauthors worked developing code for specific functions, writing vignettes, etc. However, for our simplicity, we decided to maintain a “centralized” management of the repository. Maybe not the best practice but since we were all relatively new using GitHub, PR, forking repositories, etc., we found this way as reasonable and straightforward for everyone. I hope this helps to clarify how we have managed the authorship for the package.
