+++ vignettes/available_metrics_classification.Rmd
@@ -93,18 +93,45 @@ _Note: All classification functions automatically recognize the number of classe
 
 <br/>
 
+**List of abbreviations:**
+
+P = positive (true + false)
+
+N = negative (true + false)
+
+TP = true positive
+
+TN = true negative
+
+FP = false positive
+
+FN = false negative
+
+TPR = true positive rate
+
+TNR = true negative rate
+
+FPR = false positive rate
+
+FNR = false negative rate
+
+PPV = positive predictive value
+
+B = coefficient B (a.k.a. beta) indicating the weight to be applied to the estimation of `fscore` (as $B^2$).
+
+
 ## References: <br/>
 
 01. Ting K.M. (2017). Confusion Matrix. _In: Sammut C., Webb G.I. (eds) Encyclopedia of Machine Learning and Data Mining. Springer, Boston, MA._ \doi{10.1007/978-1-4899-7687-1_50}
 
-02. (2017) Accuracy. _In: Sammut C., Webb G.I. (eds) Encyclopedia of Machine Learning and Data Mining_
+02. Accuracy. (2017). _In: Sammut C., Webb G.I. (eds) Encyclopedia of Machine Learning and Data Mining_
 _. Springer, Boston, MA._ \doi{10.1007/978-1-4899-7687-1_3}
 
 03. García, V., Mollineda, R.A., Sánchez, J.S. (2009). Index of Balanced Accuracy: A Performance Measure for Skewed Class Distributions. _In: Araujo, H., Mendonça, A.M., Pinho, A.J., Torres, M.I. (eds) Pattern Recognition and Image Analysis. IbPRIA 2009. Lecture Notes in Computer Science, vol 5524. Springer-Verlag Berlin Heidelberg._ \doi{10.1007/978-3-642-02172-5_57}
 
 04. Ting K.M. (2017). Precision and Recall. _In: Sammut C., Webb G.I. (eds) Encyclopedia of Machine Learning and Data Mining. Springer, Boston, MA._ \doi{10.1007/978-1-4899-7687-1_659}
 
-05. (2017) Sensitivity. _In: Sammut C., Webb G.I. (eds) Encyclopedia of Machine Learning and Data Mining. Springer, Boston, MA._ \doi{10.1007/978-1-4899-7687-1_751}
+05. Sensitivity. (2017). _In: Sammut C., Webb G.I. (eds) Encyclopedia of Machine Learning and Data Mining. Springer, Boston, MA._ \doi{10.1007/978-1-4899-7687-1_751}
 
 06. Ting K.M. (2017). Sensitivity and Specificity. _In: Sammut C., Webb G.I. (eds) Encyclopedia of Machine Learning and Data Mining. Springer, Boston, MA._ \doi{10.1007/978-0-387-30164-8_752}
 
