+++ _pkgdown.yml
@@ -1,6 +1,4 @@
-title: metrica
-url: https://adriancorrendo.github.io/metrica/
+url: ~
 template:
   bootstrap: 5
-  bootswatch: sandstone
 

[…]
