+++ pkgdown.yml
@@ -7,8 +7,5 @@ articles:
   available_metrics_regression: available_metrics_regression.html
   classification_case: classification_case.html
   regression_case: regression_case.html
-last_built: 2022-07-29T20:23Z
-urls:
-  reference: https://adriancorrendo.github.io/metrica/reference
-  article: https://adriancorrendo.github.io/metrica/articles
+last_built: 2022-11-02T18:44Z
 
