### 2022-11-02 63cacf0
update class metrics vignette
