### 2022-11-02 adriancorrendo/metrica#28 @adriancorrendo
Dear reviewers,

Thank you very much for your revisions that help us to improve this project!

Changes have been made to both branches of the repository (master and JOSS)

Please, you may find next the point-by-point response (one answer per comment):

> **1. Data sharing**: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
For the most part, this is true, though the data included [in some vignettes](https://adriancorrendo.github.io/metrica/articles/apsim_open.html) under the testthat directory is not immediately available for users of the software:
`soybean.out <- metrica::import_apsim_out(filepath = "../tests/testthat/examples/soybean.out")`
#...
`soybean.db <- metrica::import_apsim_db(filename = "soybean.example.db", folder = "../tests/testthat`/examples/")
I was able to read these in by pulling the source locally and adjusting the filepaths. These either ought to be included as data with the package, saved to file in the vignette, and then read in with the metrica::import_* functions, or simply read in as files from a GitHub raw URL. The file.path() function can be used to create OS-independent filepaths.

**Answer**
Thank you very much for this comment. Following the request, we have included the files with the package as part of “inst/extdata” so they are now more ready available for users. However, please note that both files have the peculiarity of not being “complete datasets” (because they only bring "predictions"). The idea of these functions is to simply import the “predictions” of the APSIM files to the R environment. Yet, users must match these imported predictions with their corresponding “observed” values (using dplyr::left_join() ) but this widely vary from one case to another, so it’s a separate step for users. Thus, any of these files can be used directly as an input to any other of the package functions to estimate prediction “performance”. They can only be input of the “import_*” functions.

- Solution
At the corresponding [vignette](https://adriancorrendo.github.io/metrica/articles/apsim_open.html), we show users how to get access to their local copy of the files, now as part of the "extdata" folder within the package directory. These lines work  in a OS-independent way.

-- For APSIM Classic (*.out)
`apsim_out_filepath <- [system.file](https://rdrr.io/r/base/system.file.html)("extdata/soybean.out", package = "metrica")`
`soybean.out <- metrica::[import_apsim_out](https://adriancorrendo.github.io/metrica/reference/import_apsim_out.html)(filepath = apsim_out_filepath)`

-- For APSIM NextGeneration (*.db)
`apsim_db_folderpath <- [system.file](https://rdrr.io/r/base/system.file.html)("extdata", package = "metrica")`
`soybean.db <- metrica::[import_apsim_db](https://adriancorrendo.github.io/metrica/reference/import_apsim_db.html)(filename = "soybean.example.db", folder = apsim_db_folderpath)`
