### 2022-11-02 adriancorrendo/metrica#28 @adriancorrendo
> **2. Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
You write in your paper "the first step is to install R (>= 4.2.0)," though your listed needed R version in DESCRIPTION is 2.1.0. These ought to be consistent, if included in the paper.

- **Answer**
Thank you for the catch! After double-checking, the correct minimum version required based on dependencies is R >= 3.6.2.
However, following comment No. 6 recommending to reduce the corresponding section, we have simply removed this line from the article. After all, being an R-package already implies that R is required.
