### 2022-11-02 adriancorrendo/metrica#28 @adriancorrendo
> **3. Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
I haven't been able to track down a CONTRIBUTING file, code of conduct, README section on contributing, or GitHub issue/PR template. I would feel confident checking off this box with any one of these present!

- **Answer**:
Thank you very much for this suggestion. As suggested, we have included:
  - [Contributing to our package](https://adriancorrendo.github.io/metrica/index.html#5-contributing-to-our-package) and [Code of conduct](https://adriancorrendo.github.io/metrica/index.html#6-code-of-conduct) sections into the README file.
   - A [CODE_OF_CONDUCT article](https://adriancorrendo.github.io/metrica/CODE_OF_CONDUCT.html) accompanying the documentation.
