### 2022-11-02 adriancorrendo/metrica#28 @adriancorrendo
> **4. Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
The help-files are really well-written. Many of the them refer to [these vignettes](https://adriancorrendo.github.io/metrica/articles/available_metrics_classification.html) giving formulas and higher-level descriptions, though many of the terms in the formulas are not documented. B^2, for instance, is ambiguous in context, and the TP/TN/FP/etc notations are not self-explanatory.

- **Answer**:
Thank you for this comment! In line with the suggestion, we have modified the [corresponding vignette](https://adriancorrendo.github.io/metrica/articles/available_metrics_classification.html#list-of-classification-metrics-categorical-variables) providing a list of abbreviations by the end of the table as follows:

**List of abbreviations:**
P = positive (true + false)
N = negative (true + false)
TP = true positive
TN = true negative
FP = false positive
FN = false negative
TPR = true positive rate
TNR = true negative rate
FPR = false positive rate
FNR = false negative rate
PPV = positive predictive value
B = coefficient B (a.k.a. beta) indicating the weight to be applied to the estimation of `fscore` (as $B^2$).
