### 2022-11-02 adriancorrendo/metrica#28 @adriancorrendo
> **5. State of the field**: Do the authors describe how this software compares to other commonly-used packages?
Your bullet points beginning at line 60 are generally informative for describing the unique advantages of your package.
In the paper, you write:
•	the most extensive collection of prediction performance metrics for regression and classification models up to date.
At first glance, this doesn't seem to be true given the packages you've cited. Loosely:
library(cvms)
length(formalArgs(gaussian_metrics)) +
length(formalArgs(binomial_metrics)) +
length(formalArgs(multinomial_metrics))
#> [1] 84
or
https://github.com/tidymodels/yardstick/blob/main/NAMESPACE#L3-L182
Metric calculation modules/libraries from non-R ecosystems, as well, can be [quite extensive](https://scikit-learn.org/stable/modules/model_evaluation.html).
...which is okay! I'd just recommend removing that line from the paper, or clarifying it with some qualifiers.

- **Answer**:
Thank you for the comments here. In line with the recommendation, and since the number of metrics of packages may evolve over time, we have softened the tone of the statement to the following at Line 60:
“- one of the most extensive collections of prediction performance metrics for regression and classification models in R.”

Please, note that we have also added that this is specifically “in R”.

Regarding the meaning of “the most extensive collection of metrics”, we would like to expand what we tried to mean. Simply, we just wanted to refer to the number of error of goodness of fit “metrics” like RMSE, MAE, etc. The counts you have provided above for cvms and yardstick is counting for “arguments” but not necessarily error “metrics”. Moreover, there are numerous classification metrics of the 84 count from cvms that are counted twice (or more) between binomial_metrics() and multinomial_metrics().

**For regression**:
1. In the case of **metrica**, the total of available regression metrics is 50:
https://adriancorrendo.github.io/metrica/articles/available_metrics_regression.html

2. **cvms** gaussian_metrics (regression) comes to a total of 20, detailed on https://cran.r-project.org/web/packages/cvms/vignettes/available_metrics.html, and given the list from gaussian_metrics():
`c(rmse,  mae ,  nrmse_rng ,  nrmse_iqr ,  nrmse_std ,  nrmse_avg ,  rae ,  rse ,  rrse ,  rmsle , male,  mape ,  mse ,  tae ,  tse ,  r2m ,  r2c ,  aic ,  aicc ,  bic)`

3. In the case of **yardstick**, detailed on https://yardstick.tidymodels.org/articles/metric-types.html, the total of regression metrics (specified as “numeric” type) are 16 given:
`c(ccc, huber_loss, huber_loss_pseudo, iic, mae, mape, mase, mpe, msd, poisson_log_loss, rmse, rpd, rpiq, rsq, rsq_trad, smape)`

**For classification**:
1. In **metrica**, the available classification metrics are a total of 26:
https://adriancorrendo.github.io/metrica/articles/available_metrics_classification.html
2. **yardstick** offers a total of 23 classification metrics as detailed here: https://yardstick.tidymodels.org/articles/metric-types.html#metrics
3. **cvms** brings a total 17 classification metrics, https://cran.r-project.org/web/packages/cvms/vignettes/available_metrics.html, which are duplicated between binomial & multinomial, but the latter bring the option to get overall, macro (one vs all), or weighted-type estimations.

Functions in **metrica**, automatically detect the number of classes to run either binomial or multinomial. Then the option of obtaining overall vs per-class metrics estimates is managed with the “atom” argument, so there is no need to repeat the function. My understanding is that **yardstick** functions do almost the same. Thus, **cvms** seems to offer a larger number of classification metrics, but actually, it’s just a repetition of the same metrics but changing the way of estimation.
