### 2022-11-02 adriancorrendo/metrica#28 @adriancorrendo
> ** 6. Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
The paper is generally well-written!
The "Package features" section is quite long, and includes elements that I haven't seen in other accepted JOSS papers. I'd recommend cutting the "Package features" section from "System requirements and installation" on, currently lines 86 through 109.
I came across a few grammatical errors while reviewing, and would like to see another round of copy edits before recommending acceptance. Would be glad to put in a PR with copy edits, if appreciated!

- **Answer**:
We deeply appreciate the suggestions and the offer to check for grammatical expressions. As recommended, we have substantially reduced the “Package features” section by removing the “Software Requirements and installation”, and the code to generate the examples **No. 1** and **No. 2**.

Nonetheless, if the reviewers agree, we would like to keep two parts of this section:
i) “Using the functions” (shortened), which provides relevant details about the basic inputs and defaults for regression, binomial and multinomial cases, and
ii) The “Examples No.1 and No. 2” which show two of the main visual outputs that can be obtained with specific package functions (but removing the code).

Regarding the offer to add copy edits, please, we would be more than glad to count with those through a PR. Thank you!
