### 2022-11-02 adriancorrendo/metrica#28 @adriancorrendo
> ** 7. References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?
Many software names in these entries are not case protected, resulting in the first letters to be capitalized / made lowercase when they ought not to be. e.g. "Cvms: ..." ought to read {cvms}:... in the LaTeX.

- **Answer**:
Thank you very much for this comment. In line with the suggestion, we have modified the citation syntax in order make the name of the packages “case protected”. We have also fixed other minor typos within the references.bib file.
