### 2022-11-02 adriancorrendo/metrica#28 @adriancorrendo
On behalf of all the collaborator of the **metrica** project, I would like to express gratitude for your time and patience revising our package.

Please, just let us know if any additional revisions are required.

Best regards,

ADRIAN
