### 2022-11-02 #4655 @adriancorrendo
> @adriancorrendo, please respond in the issue in the source repository [adriancorrendo/metrica#28](https://github.com/adriancorrendo/metrica/issues/28)
>
> The additional comment by @simonpcouch can be answered in this thread.

Dear Dr. @osorensen ,

I have just submitted the point-by-point response to revisions at the corresponding issue [https://github.com/adriancorrendo/metrica/issues/28](https://github.com/adriancorrendo/metrica/issues/28).

Please, let me know if any further step is required from my side.

Thank you for your time and patience!

Best regards,

ADRIAN
