### 2022-11-03 adriancorrendo/metrica#28 @simonpcouch
@adriancorrendo, your edits look great, and I very much appreciate the detailed responses in this thread! I'll close this issue and sign off on the remaining checkboxes in the review issue.

Yall have put together a wonderful package, and I'm glad to have had the chance to check it out.🙂
