+++ pkgdown.yml
@@ -7,5 +7,5 @@ articles:
   available_metrics_regression: available_metrics_regression.html
   classification_case: classification_case.html
   regression_case: regression_case.html
-last_built: 2022-11-02T18:44Z
+last_built: 2022-11-04T02:40Z
 
