### 2022-11-04 #4655 @osorensen
@adriancorrendo, I will now read through the paper once more, and let you know if I have any suggested changes.

In the meantime could you:
