### 2022-11-04 adriancorrendo/metrica#29 @osorensen
Editorial comments on JOSS submission
This is a comment to the submission https://github.com/openjournals/joss-reviews/issues/4655.

The links on pages 49-52 do not work, because my pdf reader seem to think that only the part on the first line is a URL.

I believe you can fix this by replacing the link in these lines

https://github.com/adriancorrendo/metrica/blob/63cacf04f5c0d9f85f41fd8f1edd6a504137ec24/paper.md?plain=1#L63-L65

with something like

```
[Regression metrics](https://adriancorrendo.github.io/metrica/articles/available_metrics_regression.html)
```

and similar for Classification metrics.
