### 2022-11-04 adriancorrendo/metrica#30 @osorensen
Editorial comments on JOSS submission
A suggested change to the JOSS submission https://github.com/openjournals/joss-reviews/issues/4655.

On line 53 you write "To extent of our knowledge ...".

I think a better English phrase here would be "To the best of our knowledge ...".
