### 2022-11-04 adriancorrendo/metrica#32 @osorensen
Editorial comments on JOSS submission
https://github.com/openjournals/joss-reviews/issues/4655

All the links that appear across multiple lines are dead. Please update them as suggested in #29.
