### 2022-11-04 baf0061
update NEWS
