### 2022-11-04 #4655 @adriancorrendo
Dear @osorensen ,

Thank you very much for your copy edits.
within the latest [commit to the JOSS branch](https://github.com/adriancorrendo/metrica/commit/3e63e81b41f7f8077642491fdc515b75413c90a2), I have made the modifications requested issues 29 to 31

[New release tag v2.0.1.999](https://github.com/adriancorrendo/metrica/releases/tag/v2.0.1.999)
- Question: Please, note the release was made on the "stable" master-branch, not on the temporary JOSS. Is it correct? Or do I need to create the new release using the JOSS-branch (containing the "paper.md" file) as the target??

[Zenodo archived version](https://zenodo.org/record/7291776)

Archived version DOI: [https://doi.org/10.5281/zenodo.7291776](https://doi.org/10.5281/zenodo.7291776)

Please, let me know if any further modification is needed from my end.

Thank you for your time!

Best regards,

ADRIAN
