### 2022-11-14 #4655 @adriancorrendo
> @editorialbot generate pdf

Good evening, Dr. @osorensen

May I ask if any action is required from my side? I see the paper has been stuck for the last 10 days after the decision

> 👋 @openjournals/dsais-eics, this paper is ready to be accepted and published.
>
> Check final proof 👉📄 [Download article](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04655/joss.04655/10.21105.joss.04655.pdf)
>
> If the paper PDF and the deposit XML files look good in [openjournals/joss-papers#3682](https://github.com/openjournals/joss-papers/pull/3682), then you can now move forward with accepting the submission by compiling again with the command `@editorialbot accept`

Thank you for your time!

Best regards,

ADRIAN
